+++ components/commercetools-ui/atoms/overlay/index.tsx
@@ -1,12 +1,17 @@
 import React from 'react';
+import useClassNames from 'helpers/hooks/useClassNames';
 
 interface Props {
+  zIndex?: number;
   onClick?: () => void;
 }
 
-const Overlay: React.FC<Props> = ({ onClick }) => {
+const Overlay: React.FC<Props> = ({ zIndex, onClick }) => {
   //eslint-disable-next-line tailwindcss/no-custom-classname
-  return <div className="fixed left-0 top-0 z-[250] h-screen w-screen bg-[#7f7f7f]/30" onClick={onClick} />;
+  const z = `z-[${zIndex ?? 600}]`;
+  return (
+    <div className={useClassNames(['fixed left-0 top-0 h-screen w-screen bg-[#7f7f7f]/30', z])} onClick={onClick} />
+  );
 };
 
 export default Overlay;