+++ components/commercetools-ui/organisms/product/product-details/components/product-variant.tsx
@@ -1,4 +1,4 @@
-import { FC, useEffect, useMemo, useState } from 'react';
+import { FC, useCallback, useMemo } from 'react';
 import { Variant as VariantType } from 'shared/types/product';
 import Typography from 'components/commercetools-ui/atoms/typography';
 import useClassNames from 'helpers/hooks/useClassNames';
@@ -25,18 +25,17 @@ const ProductVariant: FC<ProductVariantProps> = ({
 }) => {
   const { formatMessage } = useFormat({ name: 'product' });
 
-  const [variantsToUse, setVariantsToUse] = useState<VariantType[]>();
-
   const attributeString = attribute.toString();
 
   const attributeToFilterTo = attribute == 'size' ? 'colorlabel' : 'size';
 
-  useEffect(() => {
-    const filteredVariants: VariantType[] = currentVariant.attributes?.['size']
-      ? filterAttributeBasedVariants(variants, currentVariant, attributeToFilterTo)
-      : discardRepeatedValues(variants, attribute.toString());
-    setVariantsToUse(filteredVariants);
-  }, [variants, attribute, currentVariant, attributeToFilterTo]);
+  const variantsToUse: VariantType[] = useMemo(
+    () =>
+      currentVariant.attributes?.['size']
+        ? filterAttributeBasedVariants(variants, currentVariant, attributeToFilterTo)
+        : discardRepeatedValues(variants, attribute.toString()),
+    [attribute, attributeToFilterTo, currentVariant, variants],
+  );
 
   const hoverEffectClassName = useMemo(
     () => (variantsToUse && variantsToUse.length > 1 ? 'hover:cursor-pointer' : 'pointer-events-none'),
@@ -53,37 +52,40 @@ const ProductVariant: FC<ProductVariantProps> = ({
     inModalVersion ? 'text-12' : 'text-14',
   ]);
   const defaultVariantClassName = useClassNames(['h-20 w-20 rounded-full', hoverEffectClassName]);
-  const sizeVariantClassname = useClassNames(['p-12', hoverEffectClassName]);
+  const sizeVariantClassname = useClassNames(['text-14 p-12', hoverEffectClassName]);
   const variantContainerClassName = useClassNames(['mt-16 flex', attribute == 'size' ? 'gap-12' : 'gap-24']);
 
-  const sortVariantsCmp = (a: VariantType, b: VariantType) => {
-    const first = a.attributes?.[attribute];
-    const second = b.attributes?.[attribute];
+  const sortVariantsCmp = useCallback(
+    (a: VariantType, b: VariantType) => {
+      const first = a.attributes?.[attribute];
+      const second = b.attributes?.[attribute];
 
-    if (first > second) return 1;
-    if (first < second) return -1;
-    return 0;
-  };
+      if (first > second) return 1;
+      if (first < second) return -1;
+      return 0;
+    },
+    [attribute],
+  );
 
   return (
     <div className={className}>
-      <div className={attribute == 'color' ? 'mt-4' : ''}>
+      <div className={attribute == 'color' ? 'mt-4 grid gap-4' : ''}>
         <Typography className={labelClassName}>
           {formatMessage({ id: attributeString, defaultMessage: attributeString })}
         </Typography>
-        <Typography className="text-12 leading-loose">{currentVariant?.attributes?.[`${attribute}label`]}</Typography>
+        <Typography className="text-14 leading-loose">{currentVariant?.attributes?.[`${attribute}label`]}</Typography>
       </div>
 
       <div className={variantContainerClassName}>
-        {variantsToUse?.sort(sortVariantsCmp).map(({ attributes, id, sku }) => (
-          <div
+        {variantsToUse?.toSorted(sortVariantsCmp).map(({ attributes, id, sku }) => (
+          <Typography
             key={id}
+            onClick={() => onClick?.(sku)}
             className={getVariantClassName(id)}
             style={attribute !== 'size' ? { backgroundColor: attributes?.[attribute] } : {}}
-            onClick={() => onClick?.(sku)}
           >
             {attribute == 'size' && attributes?.[attribute]}
-          </div>
+          </Typography>
         ))}
       </div>
     </div>